Fixed issue on user authentication
Fixed cast issue (int -> long) that was causing errors when authenticating as a new user

+++ app/models/User.scala
@@ -149,7 +149,7 @@ object User {
           'url -> user.url,
           'bio -> user.bio,
           'location -> user.location
-        ).as(int("id").single)
+        ).as(long("id").single)
 
         //store object in cache for later retrieval
         Cache.set(userCacheKey + id, user.copy(id = Id(id)), 60*60)
